+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix newmail_notifier notification focus in Chrome (#9467)
 - Fix fatal error when parsing some TNEF attachments (#9462)
 - Fix double scrollbar when composing a mail with many plain text lines (#7760)
+- Fix decoding mail parts with multiple base64-encoded text blocks (#9290)
 
 ## Release 1.6.7